Added fPIC-fix patch, added various fixes for segfaults, added changes for VDR >= 1.5.3

+++ magazine.h
@@ -70,7 +70,7 @@ class magazine : public cOsdObject
 	enum modes {SHOW,EDIT};
 	enum modes curmode;
 
-	unsigned short EDIT_curEvent;
+	int EDIT_curEvent;
 	int EDIT_curChannel;
 	int EDIT_curEVI;
